requirements.txt
Bootstrap_Flask==2.3.3
Flask_WTF==1.2.1
WTForms==3.0.1
Werkzeug==3.0.0
Flask==2.3.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv~=1.0.1